requirements-dev.txt
-r requirements.txt
ruff==0.12.7
pytest==8.4.1
pytest-cov==6.2.1
tox==4.28.4
pre-commit==4.2.0
requests-mock==1.12.1